Move TestService to a separate file to simplify its dependency
